ft_split: split into multiple functions
